requirements.txt
python-dotenv ~= 1.0.0

hikari ~= 2.0.0.dev120
hikari-crescent ~= 0.6.2
hikari-miru ~= 3.1.2

prisma ~= 0.9.1
